fix: do not try to get properties from already deleted records
If you try to remove a record that does not exist, this threw an error
before, since `peekRecord` returns `null` if the record does not exists.
This guards against that case.

In our use case, we wanted to remove a document, but it might not exist
before and we don't want to find the record before calling remove on it.